pyproject.toml
[tool.poetry]
name = "bso_nsk"
version = "0.1.0"
description = ""
authors = ["a.zinoviev <pandarado@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
django-mptt = "^0.13.4"
django-image-cropping = "^1.7"
easy-thumbnails = "^2.8.1"
django-view-breadcrumbs = "^2.2.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
